getChannel() now returns channel to send to
getChannel() previously simply returned the channel which a message was
sent from, however this meant if a private message was sent it would
return the result of the command to itself - channel <botnick>, which is
not useful. Therefore it checks if the channel is <botnick> and if so,
it will instead return the username of the user who sent the private
message to the bot.