botnick not found fix
Botnick could not be referenced from an external file, so a new file
"settings.txt" has been made which contains "BotNick:<name of bot>".
This can be edited before startup. Potentially could be asked at launch
which edits the file, but this seems unnessasary as of now. Edited
bot.py and rawhandle.py to search for the nick.